probando excepciones

+++ src/main/java/com/example/DoloresAleman_Final/service/PacienteService.java
@@ -3,6 +3,7 @@ package com.example.DoloresAleman_Final.service;
 import com.example.DoloresAleman_Final.Model.DomicilioDTO;
 import com.example.DoloresAleman_Final.Model.PacienteDTO;
 import com.example.DoloresAleman_Final.controller.OdontologoController;
+import com.example.DoloresAleman_Final.exceptions.ResourceNotFoundException;
 import com.example.DoloresAleman_Final.persistence.entity.Domicilio;
 import com.example.DoloresAleman_Final.persistence.entity.Paciente;
 import com.example.DoloresAleman_Final.persistence.repository.IOdontologoRepository;
@@ -24,30 +25,37 @@ public class PacienteService {
     ObjectMapper mapper;
 
 
-    public String guardar(Paciente p){
+    public String guardar(Paciente p) {
         String respuesta = null;
-        if (repository.save(p) != null){
+        if (repository.save(p) != null) {
             respuesta = "ok";
         }
         return respuesta;
     }
 
-    public void eliminar(Long id){
-        repository.deleteById(id);
-    }
-
-    public Optional<Paciente> buscarPorId(Long id){
+    public Optional<Paciente> buscarPorId(Long id) {
 
         return repository.findById(id);
     }
 
-    public List<PacienteDTO> buscarTodos(){
-
-        List<PacienteDTO> pacientes = new ArrayList<>();
-
-        for(Paciente paciente: repository.findAll()){
-            pacientes.add(mapper.convertValue(paciente, PacienteDTO.class));
-        }
-        return pacientes;
+    /*public void eliminar(Long id){
+        repository.deleteById(id);
+    }*/
+    //mismo metodo pero con excepciones
+    public void eliminar(Long id) throws ResourceNotFoundException {
+        if (buscarPorId(id) == null)
+            throw new ResourceNotFoundException("no existe un paciente con id: " + id);
+        repository.deleteById(id);
     }
-}
+
+        public List<PacienteDTO> buscarTodos () {
+
+            List<PacienteDTO> pacientes = new ArrayList<>();
+
+            for (Paciente paciente : repository.findAll()) {
+                pacientes.add(mapper.convertValue(paciente, PacienteDTO.class));
+            }
+            return pacientes;
+        }
+    }
+
